Run a GC every minute in tests
